% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggPlotData.R
\name{aggPlotData}
\alias{aggPlotData}
\title{Aggregate a data frame for plotting large time-series quickly}
\usage{
aggPlotData(df_in, varNames, timeUnit)
}
\arguments{
\item{df_in}{data frame object}

\item{varNames}{aggregation method; mean, median, sum, min, max, or circular (for averaging direction measurements in degrees)}

\item{timeUnit}{aggregation window, e.g. 'days', '6 hours, 'hours'}
}
\description{
Aggregate a data frame for plotting large time-series quickly
}
\examples{
plotDF = aggPlotData(dataIn = myDF, varNames = "TmpWtr", timeUnit = 'days')

}
\keyword{plotting}
