% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggTS.R
\name{aggTS}
\alias{aggTS}
\title{Aggregate timeseries - low level function}
\usage{
aggTS(dataIn, timestep, FUN, pullAgg, outType)
}
\arguments{
\item{dataIn}{data frame object}

\item{timestep}{new timestep used in the aggregation results in sec}

\item{FUN}{aggregation method; mean, median, sum, min, max, or circular (for averaging direction measurements in degrees)}

\item{pullAgg}{aggregate data from before/on new timestamp ('left'; default), either side of timestamp ('centre'), or on/after timestamp ('right')}

\item{outType}{output data. "LF" for low resolution data frame, "HF" for original resolution data with bin, and "both" for both formats in the list}
}
\description{
Trim dates and/or aggregate data frame
}
\examples{
LF = aggTS(dataIn = myDF, timestep = 60x60x4, FUN = "mean", pullAgg = "center", outType = "LF")


}
\keyword{wrangling}
