% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyEqn.R
\name{applyEqn}
\alias{applyEqn}
\title{applyEqn}
\usage{
applyEqn(DF_in, startDate, endDate, varNames, coeffs, logID, showPlot,
  savePlot)
}
\arguments{
\item{startDate}{start date}

\item{endDate}{endDate}

\item{coeffs}{a vector of coefficients, from 0 to nth order (e.g., c(2,3,4,5) == 5x^3 + 4x^2 + 3x + 2)}

\item{rB3in}{rB3 object input}

\item{plotPath}{plot figure of before and after}
}
\description{
This custom function applies a quadratic transformation to a specifed data range
of the format newVal = oldVal*coeffA^2 + oldVal*coeffB + coeffC
}
\examples{
newDF <- DOsat2DOmg_ZebraTechDOpto(DF_in = myDF,DOmgColName = "DOconc_d00050", DOsatColName = "DOpsat_d00050", TColName = "TmpDOs_d00050")


}
\keyword{customFun}
