% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyInterp.R
\name{applyInterp}
\alias{applyInterp}
\title{Interpolate NA values}
\usage{
applyInterp(rB3in, startDate, endDate, varNames, maxNArep, logID, Reason,
  showPlot, savePlot)
}
\arguments{
\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{variable names or keywords}

\item{maxNArep}{maximum repeated NA}

\item{logID}{assign log ID}

\item{plotPath}{plot figure of before and after}
}
\description{
Assign linearly interpolated numbers for na when na is repeated less than maxNArep number of times.
}
\examples{
rB3out <- applyInterp(rB3in, maxNArep = 4)

}
\keyword{QA/QC}
