% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyNth.R
\name{applyNth}
\alias{applyNth}
\title{Apply an nth order transfomration}
\usage{
applyNth(rB3in, startDate, endDate, varNames, coeffs, logID, Reason,
  showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{coeffs}{a vector of coefficients, from 0 to nth order (e.g., c(2,3,4,5) == 5x^3 + 4x^2 + 3x + 2)}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
Apply a mathematical transformation, e.g. new = a + b(old) + c(old)^2 + c(old)^3 + ...etc
}
\examples{
rB3in <- applyEqn(rB3in = rB3in,varNames = 'TmpWtr', coeffs = c(1,1,1,1), showPlot = TRUE)


}
\keyword{transformations}
