% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignInterp.R
\name{assignInterp}
\alias{assignInterp}
\title{Interpolate the date range given for the variables specified}
\usage{
assignInterp(rB3in, startDate, endDate, varNames, logID, Reason, showPlot,
  savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{logID}{write an operation identifier to the log frames, default = NA, 'auto' chooses a log number for you, or provide a numerical code}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
Linearly interpolates gaps (NAs) within the selected subset of dates and variables
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
