% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bobs2df.R
\name{bobs2df}
\alias{bobs2df}
\title{Buoy observation to data frame}
\usage{
bobs2df(filePath, metaD, useHydroYear)
}
\arguments{
\item{filePath}{file path}

\item{metaD}{meta data}

\item{useHydroYear}{expand dataset to hydroyear limits}
}
\description{
Import csv data into data frame
}
\examples{
myDF <- bobs2df("Rotorua.bobs",metaD, useHydroYear = TRUE)


}
\keyword{fileIO}
