% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cachedir}
\alias{cachedir}
\title{Returns rb3 package cache directory}
\usage{
cachedir()
}
\value{
a string with the file path of rb3 cache directory
}
\description{
Returns rb3 package cache directory
}
\examples{
cachedir()
}
