% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-cdi.R
\name{cdi-idi}
\alias{cdi-idi}
\alias{cdi_get}
\alias{idi_get}
\title{Get CDI rate and IDI index value from B3 front page}
\usage{
cdi_get()

idi_get()
}
\value{
\code{data.frame} with CDI rate or IDI index values.
}
\description{
Scrape page \url{https://www.b3.com.br/} to get last available CDI rate and
IDI index values.
}
\examples{
df <- cdi_get()
df <- idi_get()
}
