% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-cotahist.R
\name{cotahist_get}
\alias{cotahist_get}
\title{Get COTAHIST data from B3}
\usage{
cotahist_get(year)
}
\arguments{
\item{year}{the year of the COTAHIST file.

All valueable information is in the \code{HistoricalPrices} element of the
returned list.
\code{Header} and \code{Trailer} have informations regarding file generation.}
}
\value{
a list with 3 data.frames: \code{Header}, \code{HistoricalPrices}, \code{Trailer}.
}
\description{
Download COTAHIST file and parses it returning structured data into R
objects.
}
\examples{
\dontrun{
df <- cotahist_get(2001)
}

}
