% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2rB3.R
\name{csv2rB3}
\alias{csv2rB3}
\title{Buoy observation to data frame}
\usage{
csv2rB3(filePath, siteName, lat, lon, country, copySrc)
}
\arguments{
\item{filePath}{file path of source csv file, timestamps in first column as yyyy-mm-dd hh:mm:ss with header "DateTime"}

\item{siteName}{name of the monitoring station (character)}

\item{lat}{latitude of the monitoring station (numeric, decimal WGS84 degrees)}

\item{lon}{longitude of the monitoring station (numeric, decimal WGS84 degrees)}

\item{country}{country of origin (character)}

\item{copySrc}{keep a copy of the raw (input) file as a dataframe (TRUE/FALSE) (stored as 'srcDF')}
}
\description{
Imports a csv file to rB3
}
\details{
Creates a list of dataframes, including: \cr
     1. the unmodified data ('srcDF')\cr
     2. a copy of the data, which will be modified by rB3 ('qcDF')\cr
     3. a 'log' file with similar structure to srcDF and qcDF, to store QC operations log ('logDF')\cr
     4. controls, extracted from teh header rows in the raw csv files ('ctrls')\cr
     5. site metadata, create from additional input args ('metaD')\cr
}
\examples{
csv2rB3("Rotorua.csv", siteName = "Lake_Rotorua", lat = -38, long = 176, country = "NZ")
}
\keyword{fileIO}
