% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destructObj.R
\name{destructObj}
\alias{destructObj}
\title{Decompose object structure and export desired content}
\usage{
destructObj(dataIn, exportName, exportOpt)
}
\arguments{
\item{exportName}{default: qcDF}

\item{exportOpt}{default: 1.
1: return specified content.
2. return specified content and object type (1: rB3obj, 2 = data.frame)}
}
\description{
Decompose object structure and export desired content
}
\examples{
LF = aggTS(dataIn = myData)


}
\keyword{system}
