% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-display-template.R
\name{display_template}
\alias{display_template}
\title{Display templates}
\usage{
display_template()
}
\value{
Addin has no return
}
\description{
\code{display_template} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{http://rstudio.github.io/rstudioaddins/}{addin} that allows users
to query for specific attributes of templates.
}
\examples{
\dontrun{
display_template()
}
}
