% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destructObj.R
\name{.destructObj}
\alias{.destructObj}
\title{Get qc data - private functions}
\usage{
.destructObj(dataIn, exportName, exportOpt)
}
\arguments{
\item{dataIn}{data input}

\item{exportName}{default: qcDF}

\item{exportOpt}{default: 1.
1: return specified content.
2. return specified content and object type (1: rB3obj, 2 = data.frame)}
}
\description{
Export qc data regardless of rB3obj or data.table
}
\examples{
myData = destructObj(rB3data)
myData = destructObj(myDF)
objType = unlist(destructObj(rB3data,exportOpt = 2)[2])

}
\keyword{system}
