% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_data}
\alias{download_data}
\title{Download datasets}
\usage{
download_data(template, dest = NULL, overwrite = TRUE, ...)
}
\arguments{
\item{template}{the template name}

\item{dest}{a string which specifies a valid directory or file name to save
the file. If it is not provided a temporary filename is generated.}

\item{overwrite}{a logical indicating if the file must be overwritten once
\code{dest} already exists.}

\item{...}{aditional arguments}
}
\value{
a string with the file path of downloaded file or \code{NULL} if download
fails.

This function downloads data sets for those templates that specifies a
\code{downloader} attribute.
}
\description{
Download datasets for a given template.
}
\examples{
\dontrun{
fname <- download_data("CDIIDI")
}

}
