% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftCorr.R
\name{driftCorr}
\alias{driftCorr}
\title{Apply drift correction}
\usage{
driftCorr(rB3in, startDate, endDate, varNames, lowRef, lowStart, lowEnd,
  highRef, highStart, highEnd, logID, Reason, showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{variable to which drift correction will be applied}

\item{lowRef}{a reference/known low value at start of period}

\item{lowStart}{equivalent sensor value at start of period which should = lowRef  (i.e., starting offset)}

\item{lowEnd}{equivalent sensor value at end of period which should = lowRef  (i.e., finishing offset)}

\item{highRef}{a reference/known high value at start of period}

\item{highStart}{equivalent sensor value at start of period which should = lowRef  (i.e., starting offset)}

\item{highEnd}{equivalent sensor value at end of period which should = lowRef  (i.e., finishing offset)}

\item{logID}{assign log ID}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}

\item{logID}{write an operation identifier to the log frames, default = NA}
}
\description{
Change offset and slope over time to account for linear measurement drift
}
\examples{
rB3in <- driftCorr(rB3in = myDF,varNames = "DOpsat.d00050",lowRef = 0, lowStart = -10,lowEnd = 0, highRef = 100, highStart = 85, highEnd = 130, showPlot = TRUE)


}
\keyword{transformations}
