% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_vars.R
\name{exclude_vars}
\alias{exclude_vars}
\title{Specify what variables to exclude}
\usage{
exclude_vars(DF_in, metaD, varNames)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{varNames}{list of variable names or keywords}
}
\description{
This function specifies variables to be excluded for later changes.
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))

}
\keyword{wrangling}
