% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastStdze.R
\name{fastStdze}
\alias{fastStdze}
\title{Buoy observation to data frame}
\usage{
fastStdze(rB3in, startDate, endDate, varNames, timestep, methodAgg,
  pullAgg)
}
\arguments{
\item{rB3in}{rB3 object to be 'standardized'}

\item{startDate}{measurements prior to this time will be removed}

\item{endDate}{measurements after this time will be removed}

\item{varNames}{variables (columns) to be retained, select by full header name, key characters, or 'All'}

\item{timestep}{standardise (aggregate) to a common timestep (T/F; turn aggregation on or off)}

\item{methodAgg}{aggregation method; mean, median, sum, min, max, or circular (for averaging direction measurements in degrees)}

\item{pullAgg}{aggregate data from before/on new timestamp ('left'), either side of timestamp ('centre'), or on/after timestamp ('right')}
}
\description{
Trim dates and/or aggregate (to a single common timestep) an rB3 working dataframe.
    Use 'ctrls' to define how aggregation should be performed for each time-series(column).
}
\examples{
stdDF <- rB3stdze(rB3in = rawDF, varNames = wqVars, metaD = metaD, startDate = '2016-07-01 00:00:00', endDate = '2018-06-30 23:45:00', doAgg = TRUE, methodAgg = ctrls$methodAgg, pullAgg = ctrls$pullAgg)


}
\keyword{fileIO}
