% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_max.R
\name{filter_max}
\alias{filter_max}
\title{Assign NA to the specified dates range and variables}
\usage{
filter_max(DF_in, metaD, maxVal, varNames, startDate, endDate, cndFile, logID)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{endDate}
}
\description{
This function assign NA to the specific dates range and variables
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
