% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_roc.R
\name{filter_roc}
\alias{filter_roc}
\title{Find values' rate of change exceeded maxRoc value and assign NA}
\usage{
filter_roc(DF_in, metaD, maxRoc, varNames, startDate, endDate, cndFile, logID)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{maxRoc}{rate of change per timestep}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{end date}

\item{cndFile}{condition file path}
}
\description{
This function assign NA to the specific dates range and variables
}
\examples{
newDF <- maxRoc(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
