% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_stall.R
\name{filter_stall}
\alias{filter_stall}
\title{Remove repeated values}
\usage{
filter_stall(DF_in, metaD, maxRep, varNames, startDate, endDate, cndFile, logID)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{cndFile}{conditoin file path}

\item{logID}{assign log ID}

\item{minVal}{minimum value permitted}
}
\description{
This function assign NA to the repeated data more than maxRep times continuous
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
