% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMeanSD.R
\name{filterMeanSD}
\alias{filterMeanSD}
\title{Remove values that exceed rate of change}
\usage{
filterMeanSD(rB3in, maxRoc, varNames, startDate, endDate, logID, showPlot,
  savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{maxRoc}{maximum rate of change value permitted}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{logID}{assign log ID}

\item{showPlot}{plot figure of before and after}

\item{savePlot}{plot figure of before and after}
}
\description{
This function assign NA to the data larger than the assigned maxRoc
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
