% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMinMax.R
\name{filterMinMax}
\alias{filterMinMax}
\title{Assign NA based on filter values from the "ctrls" data frame}
\usage{
filterMinMax(rB3in, startDate, endDate, varNames, filterMin, filterMax, logID,
  showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{filterMin}{apply min value filter, either numeric (filterMin = <value>), or from 'crtls' (filterMin = TRUE)}

\item{filterMax}{apply max value filter, either numeric (filterMax = <value>), or from 'crtls' (filterMax = TRUE)}

\item{logID}{write an operation identifier to the log frames, default = NA}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
Assigns NA to the selected subset of dates and variables
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
