% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMSD.R
\name{filterMsd}
\alias{filterMsd}
\title{Detect and remove outliers based on running mean with standard deviation}
\usage{
filterMsd(rB3in, startDate, endDate, varNames, avgWindow, nSD, logID, showPlot,
  savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{avgWindow}{time in hrs to calculate running mean and SD}

\item{nSD}{number of stdard deviations about mean which is considered an outlier}

\item{logID}{write an operation identifier to the log frames, default = NA}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
Specify a time window, and number of SD allowed manually, or read from the 'ctrls' (which have been read from source dataset headers)
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
