% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterReps.R
\name{filterReps}
\alias{filterReps}
\title{Remove consecutive repeated values}
\usage{
filterReps(rB3in, startDate, endDate, varNames, maxReps, logID, Reason,
  showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{maxReps}{apply maximum allowed consecutive identical values filter, numerical or TRUE for from 'ctrls'}

\item{logID}{write an operation identifier to the log frames}

\item{Reason}{reason for hte logged changes (character)}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
Specify maximum number of allowable consecutive identical values, or read from the 'ctrls' (which have been read from source dataset headers)
}
\examples{
new <- filterReps(old, varNames = 'TmpWtr', maxReps = 20, showPlot = TRUE)
}
\keyword{wrangling}
