% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRoc.R
\name{filterRoc}
\alias{filterRoc}
\title{Remove values that exceed rate of change - replace with NA}
\usage{
filterRoc(rB3in, startDate, endDate, varNames, maxRoc, logID, Reason,
  showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{maxRoc}{maximum allowed rate of change between consecutive values (irrespective of time diff)}

\item{logID}{write an operation identifier to the log frames, default = NA}

\item{showPlot}{display figure in plots window, and prompt to accept changes (TRUE/FALSE)}

\item{savePlot}{save the figure to a path (provide path relative to working directory)}
}
\description{
Specify a maximum ROC manually, or read from the 'ctrls' (which have been read from source dataset headers)
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{outliers}
