% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_drift.R
\name{fix_drift}
\alias{fix_drift}
\title{Fix linear drift (for sensor maintenance calibration)}
\usage{
fix_drift(DF_in, metaD, startDate, endDate, varNames, colNum, offsets, spans,
  logID, plotPath)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{variable names or keywords}

\item{colNum}{variable column numbers}

\item{offsets}{pair of offset values corresponding to startDate and endDate (i.e. lower limits)}

\item{spans}{pair of spans values corresponding to startDate and endDate (i.e. value ranges from offset; upper limits - lower limits)}

\item{plotPath}{plot paths with name (e.g. figures/Rotorua_)}
}
\description{
Linear drift correction. First offset and span will be assumed to be the right values.
If span is missing, only offset will be used.
}
\details{
new_value(t) = (value(t) - offset(t)) / span(t) * initial_span + initial_offset
}
\examples{
newDF <- fix_drift(DF_in = myDF, startDate = "2007/1/1", endDate = "2007/7/1", colNum = 15, offsets = c(1,3), span = c(5,6), logID = 5, plotPath = "figures/Rotorua_")

}
\keyword{customFun}
