% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_template_1.R
\name{fun_template_1}
\alias{fun_template_1}
\title{Custom fun: template 1}
\usage{
fun_template_1(DF_in, metaD, startDate, endDate, varNames, colNum, logID,
  plotPath)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{variable names or keywords}

\item{logID}{assign log ID}

\item{plotPath}{plot figure of before and after}
}
\description{
This custom function converts DOsat and temperature to DOmg
}
\examples{
newDF <- custom_fun_template_1(DF_in,metaD,startDate,endDate,varNames,logID,plotPath)

}
\keyword{customFun}
