% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCrB3.R
\name{FUNCrB3}
\alias{FUNCrB3}
\title{Apply a custom (user) function}
\usage{
FUNCrB3(rB3in, startDate, endDate, varNames, FUN, eqnVars, logID, Reason,
  showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}

\item{eqnVarnames}{a vector of variables (!must call each by name) that are used in the custom function}
}
\description{
Define a custom function and use via rB3 (apply logs etc)
}
\examples{
rB3in <- applyEqn(rB3in = rB3in,varNames = 'TmpWtr', coeffs = c(1,1,1,1), showPlot = TRUE)


}
\keyword{transformations}
