% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idElToModify.R
\name{idElToModify}
\alias{idElToModify}
\title{Identify elements locations for the specified variable and dates ranges}
\usage{
idElToModify(rB3in, startDate, endDate, varNames)
}
\arguments{
\item{rB3in}{data frame input}

\item{startDate}{start date}

\item{endDate}{end date}

\item{varNames}{variable names that will be selected}
}
\description{
This sub-function identifies elements of the data frame that satisfies dates range and/or variable names,
}
\examples{
idElToModify(rB3in, startDate = '2015-07-01', endDate = '2016-06-30', varNames = 'All')

}
\keyword{wrangling}
