% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_dates.R
\name{include_dates}
\alias{include_dates}
\title{include_dates - define date ranges}
\usage{
include_dates(DF_in, metaD, startDate, endDate)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{startDate}{start date}

\item{endDate}{endDate}
}
\description{
Specify dates to include in the later tasks. Outside of these dates are removed from the data
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
