% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_vars.R
\name{include_vars}
\alias{include_vars}
\title{Specify what variables to include}
\usage{
include_vars(DF_in, metaD, varNames)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{varNames}{list of variable names or keywords}
}
\description{
This function specifies variables to be included for later changes.
}
\examples{
newDF <- include_vars(myDF,metaData,varNames = c("tmp","do","wnd","pH"))

}
\keyword{wrangling}
