% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-indexes.R
\name{index_by_segment_get}
\alias{index_by_segment_get}
\title{Get B3 indexes available}
\usage{
index_by_segment_get(index_name, cache_folder = cachedir(), do_cache = TRUE)
}
\arguments{
\item{index_name}{a string with the index name}

\item{cache_folder}{Location of cache folder (default = cachedir())}

\item{do_cache}{Whether to use cache or not (default = TRUE)}
}
\value{
A dataframe with the index stocks, their weights, segments and positions.
}
\description{
Gets B3 indexes available.
}
\examples{
\dontrun{
index_by_segment_get("IBOV")
}
}
