% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-indexes.R
\name{index_comp_get}
\alias{index_comp_get}
\title{Get composition of B3 indexes}
\usage{
index_comp_get(index_name, cache_folder = cachedir(), do_cache = TRUE)
}
\arguments{
\item{index_name}{a string with the index name}

\item{cache_folder}{Location of cache folder (default = cachedir())}

\item{do_cache}{Whether to use cache or not (default = TRUE)}
}
\value{
a character vector with symbols that belong to the given index name
}
\description{
Gets the composition of listed B3 indexes.
}
\examples{
index_comp_get("IBOV")
}
