% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-indexes.R
\name{index_get}
\alias{index_get}
\title{Get index historical data}
\usage{
index_get(
  index_name,
  first_date,
  last_date = Sys.Date(),
  cache_folder = cachedir(),
  do_cache = TRUE
)
}
\arguments{
\item{index_name}{a string with the index name}

\item{first_date}{First date}

\item{last_date}{Last date}

\item{cache_folder}{Location of cache folder (default = cachedir())}

\item{do_cache}{Whether to use cache or not (default = TRUE)}
}
\value{
A data.frame/tibble with index data
}
\description{
Gets historical data from B3 indexes
}
\examples{
\dontrun{
index_get("IBOV", as.Date("1977-01-01"), as.Date("1999-12-31"))
}

}
