% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-indexes.R
\name{index_weights_get}
\alias{index_weights_get}
\title{Get the assets weights of B3 indexes}
\usage{
index_weights_get(index_name, cache_folder = cachedir(), do_cache = TRUE)
}
\arguments{
\item{index_name}{a string with the index name}

\item{cache_folder}{Location of cache folder (default = cachedir())}

\item{do_cache}{Whether to use cache or not (default = TRUE)}
}
\value{
data.frame with symbols that belong to the given index name with its weights
and theorical positions.
}
\description{
Gets the assets weights of B3 indexes.
}
\examples{
\dontrun{
index_weights_get("IBOV")
}
}
