% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-indexes.R
\name{indexes_composition_get}
\alias{indexes_composition_get}
\title{Retrieve Composition of B3 Indexes}
\usage{
indexes_composition_get()
}
\value{
A data frame containing the columns:
\describe{
\item{update_date}{The date when the data was last updated.}
\item{symbol}{The symbol of the asset.}
\item{indexes}{The indexes associated with the asset.}
}

An \code{arrow_dplyr_query} or \code{ArrowObject}, representing a lazily evaluated query. The underlying data is not
collected until explicitly requested, allowing efficient manipulation of large datasets without immediate
memory usage.
To trigger evaluation and return the results as an R \code{tibble}, use \code{collect()}.
}
\description{
This function fetches the composition of B3 indexes.
It uses the template dataset "b3-indexes-composition" to retrieve the data.
}
\examples{
\dontrun{
  indexes_composition <- indexes_composition_get()
  head(indexes_composition)
}
}
