% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-indexes.R
\name{indexes_last_update}
\alias{indexes_last_update}
\title{Get the date of indexes composition last update}
\usage{
indexes_last_update(cache_folder = cachedir(), do_cache = TRUE)
}
\arguments{
\item{cache_folder}{Location of cache folder (default = cachedir())}

\item{do_cache}{Whether to use cache or not (default = TRUE)}
}
\value{
the Date when the indexes have been updated
}
\description{
Gets the date where the indexes have been updated lastly.
}
\examples{
\dontrun{
indexes_last_update()
}
}
