% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_na.R
\name{interp_na}
\alias{interp_na}
\title{Interpolate na values}
\usage{
interp_na(DF_in, metaD, startDate, endDate, varNames, logID, plotPath, maxRep)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{variable names or keywords}

\item{logID}{assign log ID}

\item{plotPath}{plot figure of before and after}

\item{maxRep}{maximum repeated NA}
}
\description{
Assign interpolated numbers for na when na is repeated less than maxRep number of time.
}
\examples{
newDF <- custom_fun_template_1(DF_in,metaD,startDate,endDate,varNames,logID,plotPath)

}
\keyword{customFun}
