% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marketdata.R
\docType{package}
\name{marketdataBR}
\alias{marketdataBR}
\alias{marketdataBR-package}
\title{Read files from Brazilian Financial Market}
\description{
Read the many files used in Brazilian Financial Market and
convert them into useful formats and data structures.
}

