% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkLongLog.R
\name{mkLongLog}
\alias{mkLongLog}
\title{Assign NA to the specified dates range and variables}
\usage{
mkLongLog(inLog, thisLog, logID)
}
\description{
This function assign NA to the specific dates range and variables
}
