% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiff.R
\name{plotDiff}
\alias{plotDiff}
\title{Sub-module to plot two timeseries affected by the changes}
\usage{
plotDiff(DF_pre, DF_post, varNames, colNum, plotPath, custom_dpi, taskName,
  isScatter)
}
\arguments{
\item{DF_pre}{data frame 1}

\item{DF_post}{data frame 2}

\item{varNames}{list of variable names or keywords to plot}

\item{colNum}{column number to plot}

\item{plotPath}{path of the images to be saved}

\item{isScatter}{if TRUE, plot scatter with lines}
}
\description{
Sub-module to plot two timeseries (called from modules)
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{fileIO}
