% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prePostPlot.R
\name{prePostPlot}
\alias{prePostPlot}
\title{gg_facetVar}
\usage{
prePostPlot(rB3in, startDate, endDate, varNames, plotLabels, srcColour,
  preColour, qcColour, showPlot, savePlot, dpi)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{measurements before this will be excluded from plot}

\item{endDate}{measurements after this will be excluded from plot}

\item{varNames}{variable names or keywords}

\item{plotLabels}{vector of titles for each panel NB: n and order of plotLabels must be same as columns for rB3in}

\item{srcColour}{colour of the unmodified data (leave out to plot only quality controlled data)}

\item{qcColour}{colour of the quality controlled data (leave out to plot only unmodified/raw data)}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to this path ('folder/start_of_file_name')}
}
\description{
This function plots variables on a faceted ggplot
}
\examples{
newDF <- gg_facetVar(rB3in = stdDF, metaD = metaD, varNames = wqVars, srcColour = 'red',
              qcColour = 'black', savePlot = 'figures/RAW_WQ_',  dpi = 400)

}
\keyword{customFun}
