% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb3-package.R
\docType{package}
\name{rb3}
\alias{rb3}
\title{Read files from Brazilian Financial Market}
\description{
Read the many files used in Brazilian Financial Market and
convert them into useful formats and data structures.
}
