% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rB3getVars.R
\name{rB3getVars}
\alias{rB3getVars}
\title{Return working variable names for the dataset}
\usage{
rB3getVars(rB3in, varNames)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{varNames}{list of variable names or keywords}
}
\description{
Return working variable names for the dataset
}
\examples{
rB3getVars(rB3in = rB3in, varNames = 'TmpWtr')


}
\keyword{data}
