% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rB3plotr.R
\name{rB3plotr}
\alias{rB3plotr}
\title{Produce a standard rB3 format ggplot}
\usage{
rB3plotr(df_in, siteName, dateStart, dateEnd, colours, facet, showPlot,
  savePlot, dpi)
}
\arguments{
\item{df_in}{data frame input, must be long format with DateTime, var, value and src}

\item{siteName}{name of site for saving plot}

\item{dateStart}{start date}

\item{dateEnd}{end date}

\item{colours}{vector of colours to use when plotting (qc then raw), or choose 'auto'}

\item{facet}{facet the plot by variable (TRUE/FALSE)}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to this path ('folder/start_of_file_name')}

\item{dpi}{dpts per inch for saved plot}
}
\description{
Produce a standard rB3 format ggplot
}
\examples{
rB3plotr(df_in, dateStart = '2015-07-01', dateEnd = '2016-06-30')

}
\keyword{plotting}
