% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rB3stdze.R
\name{rB3stdze}
\alias{rB3stdze}
\title{Buoy observation to data frame}
\usage{
rB3stdze(rB3in, startDate, endDate, varNames, timestep, aggAll, methodAgg,
  pullAgg)
}
\arguments{
\item{rB3in}{rB3 object to be 'standardized'}

\item{startDate}{measurements prior to this time will be removed}

\item{endDate}{measurements after this time will be removed}

\item{varNames}{variables (columns) to be retained, select by full header name, key characters, or 'All'}

\item{timestep}{standardise (aggregate) to a common timestep (numeric, in mins, to turn aggregation on) !! NOTE! logs will be reset to NA!!}

\item{aggAll}{TRUE: aggregate source and qc data separately, FALSE: source data = qc data.}

\item{methodAgg}{aggregation method; mean, median, sum, min, max, or circular (for averaging direction measurements in degrees)}

\item{pullAgg}{aggregate data from before/on new timestamp ('left'), either side of timestamp ('centre'), or on/after timestamp ('right')}
}
\description{
Trim dates and/or aggregate (to a single common timestep) an rB3 working dataframe.
    Use 'ctrls' to define how aggregation should be performed for each time-series(column).
}
\examples{
stdDF <- rB3stdze(rB3in = rawDF, varNames = wqVars, metaD = metaD, startDate = '2016-07-01 00:00:00', endDate = '2018-06-30 23:45:00', doAgg = TRUE, methodAgg = ctrls$methodAgg, pullAgg = ctrls$pullAgg)


}
\keyword{fileIO}
