% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCndFile.R
\name{readCndFile}
\alias{readCndFile}
\title{read .cnd file}
\usage{
readCndFile(condFilePath, rowNameList)
}
\arguments{
\item{condFilePath}{condition file path (filename)}

\item{rowNameList}{what variables to read. c("Variables",...)}
}
\description{
This sub-function read .cnd file for different condition settings for each variables.
}
