% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLong.R
\name{readLong}
\alias{readLong}
\title{read csv with colums(dateTime, depth, varIDs, values) and return list of unique varIDs}
\usage{
readLong(filePath)
}
\arguments{
\item{filePath}{file path of source csv file, timestamps in first column as yyyy-mm-dd hh:mm:ss with header "DateTime"}
}
\description{
read csv with colums(dateTime, depth, varIDs, values) and return list of unique varIDs
}
\examples{
readLong("Rotorua.csv")
}
\keyword{fileIO}
