% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyrB3.R
\name{shinyrB3}
\alias{shinyrB3}
\title{Interactive data visualisation}
\usage{
shinyrB3(rB3in, startDate, endDate)
}
\arguments{
\item{rB3in}{rB3 object to be displayed}

\item{startDate}{start date for plot}

\item{endDate}{end date for plot}

\item{varNames}{name of the (single!) variable to be plotted}

\item{colNum}{location of the column, alternative to varNames}

\item{srcColour}{colour of the unmodified data (leave out to plot only quality controlled data)}

\item{qcColour}{colour of the quality controlled data (leave out to plot only unmodified/raw data)}
}
\description{
Plot single timeseries of source and quality controlled data. Can zoom in (dbl-click). \cr
  Return x and y information in response to clicks. \cr
  Example code is generated for removing (set to NA) the highlighted area of the plot
}
\examples{
shinySetOut <- shinySet(newDF, varNames = "TmpWtr.d00500", endDate = '2018-07-01')

}
\keyword{editing}
\keyword{visual}
