% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySet.R
\name{shinySet}
\alias{shinySet}
\title{shiny plot - visually assist data transformation information}
\usage{
shinySet(DF_in, startDate, endDate, varNames, colNum)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{name of the variable}

\item{colNum}{location of the column}
}
\description{
Plot single timeseries and return x and y informatoin in response to clicks. Use ggplot. Can zoom in.
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))

}
\keyword{wrangling}
