% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySet_nogg.R
\name{shinySet_nogg}
\alias{shinySet_nogg}
\title{shiny plot - visually assist data transformation information}
\usage{
shinySet_nogg(DF_in, startDate, endDate, varNames, colNum)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{name of the variable}

\item{colNum}{location of the column}
}
\description{
Plot single timeseries and return x and y informatoin in response to clicks. No ggplot. No zooming.
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))

}
\keyword{wrangling}
