% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySet_gg.R
\name{shinyVar}
\alias{shinyVar}
\title{shinyVar - visually assist data transformation process                         {rB3}}
\usage{
shinyVar(rB3in, startDate, endDate, varNames, colNum, srcColour, qcColour)
}
\arguments{
\item{rB3in}{rB3 object to be displayed}

\item{startDate}{start date for plot}

\item{endDate}{end date for plot}

\item{varNames}{name of the (single!) variable to be plotted}

\item{colNum}{location of the column, alternative to varNames}

\item{srcColour}{colour of the unmodified data (leave out to plot only quality controlled data)}

\item{qcColour}{colour of the quality controlled data (leave out to plot only unmodified/raw data)}
}
\description{
Plot single timeseries of source and quality controlled data. Can zoom in (dbl-click).
  Return x and y information in response to clicks.
  Exmaple code is generated for removing (set to NA) the highlighted area of the pot
}
\examples{
shinySetOut <- shinySet(newDF, varNames = "TmpWtr.d00500", endDate = '2018-07-01')

}
\keyword{editing}
\keyword{visual}
