% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyView.R
\name{shinyView}
\alias{shinyView}
\title{include_dates - define date ranges}
\usage{
shinyView(DF_in, startDate, endDate, varNames)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}
}
\description{
Specify dates to include in the later tasks. Outside of these dates are removed from the data
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))

}
\keyword{wrangling}
