% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-show-templates.R
\name{show_templates}
\alias{show_templates}
\title{Show templates.}
\usage{
show_templates()
}
\value{
Addin has no return
}
\description{
\code{display_template} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{http://rstudio.github.io/rstudioaddins/}{addin} that allows users
to view the available templates.
}
