% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseLR.R
\name{sparseLR}
\alias{sparseLR}
\title{read a csv file with colums(dateTime, depth, varIDs, values) and merge exisiting HR dataset}
\usage{
sparseLR(dataIn, filePath, varNames, suffix, sparseMethod, minZ, maxZ)
}
\arguments{
\item{dataIn}{pre-existing data input; either data frame or rB3 object}

\item{filePath}{file path of the low resolution csv file}

\item{varNames}{variable names of low resolution input that you want to import}

\item{suffix}{suffix you want to add to the low resolution input varNames}

\item{sparseMethod}{method of interpolation}

\item{minZ}{minimum depth to be averaged}

\item{maxZ}{mazximum depth to be averaged}
}
\description{
read a csv file with colums(dateTime, depth, varIDs, values) and merge exisiting HR dataset
}
\examples{
sparseLR(dataIn = rB3in, filePath = "FLD_LakeRotorua.csv")
sparseLR(dataIn = rB3in, filePath = "FLD_LakeRotorua.csv", varNames = "DO", suffix = "_FLD", minZ = 0, maxZ = 5)

}
\keyword{fileIO}
