% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-cotahist.R, R/scraper-yc.R
\name{superdataset}
\alias{superdataset}
\alias{cotahist_get_options_by_symbols}
\alias{yc_brl_with_futures}
\alias{yc_usd_with_futures}
\alias{yc_ipca_with_futures}
\title{Enhanced Dataset Creation}
\usage{
cotahist_get_options_by_symbols(symbols)

yc_brl_with_futures(refdate)

yc_usd_with_futures(refdate)

yc_ipca_with_futures(refdate)
}
\arguments{
\item{symbols}{list of symbols to extract market data from the COTAHIST dataset.}
}
\value{
The function \code{cotahist_options_by_symbol_superset()} return an object that inherits from a \code{arrow_dplyr_query}
since it tries to preserve the lazy evaluation and avoids collecting the data before its return.

The functions \code{yc_brl_with_futures()}, \code{yc_usd_with_futures()}, and \code{yc_ipca_with_futures()} return an object
that inherits from the \code{data.frame} class. In cases where certain operations cannot be executed using Arrow
operators, the data must be collected before being returned.
}
\description{
To maximize the utility of B3's existing datasets, several functions integrate data from multiple
sources to generate specialized datasets for specific analytical needs. For instance,
\code{cotahist_equity_options_superset()} combines data from COTAHIST datasets
(\code{b3-cotahist-yearly}, \code{b3-cotahist-monthly}, and \code{b3-cotahist-daily}) and Reference Rates
(\code{b3-reference-rates}) to construct a dataset containing stock options data. This dataset
includes details such as the closing price of the underlying stock, its ticker symbol, and the
applicable interest rate at option expiration. This comprehensive data enables users to perform
option pricing and calculate implied volatility.
}
\details{
The functions \code{cotahist_equity_options_superset()}, \code{cotahist_funds_options_superset()},
\code{cotahist_index_options_superset()}, and \code{cotahist_options_by_symbol_superset()} use
information from the COTAHIST datasets (\code{b3-cotahist-yearly}, \code{b3-cotahist-monthly},
and \code{b3-cotahist-daily}) and Reference Rates (\code{b3-reference-rates}) and return a dataframe
containing stock option data, including the closing price of the underlying stocks, the ticker
of the underlying asset, and the interest rate at the option's expiration. The returned dataframe
contains the following columns: "refdate", "symbol", "type", "symbol_underlying",
"strike_price", "maturity_date", "r_252", "close", "close_underlying", "volume",
"trade_quantity", and "traded_contracts".

\code{cotahist_options_by_symbol_superset()} returns the same dataset but filtered for the specified asset ticker.

Returned objects preserve lazy evaluation whenever possible and avoid being
collected until the last possible moment. Exceptions occur when operations
cannot be performed using Arrow's operators — in such cases, data will be
collected and \code{data.frame}s will be returned. Please refer to the documentation
to identify the situations where this behavior applies.

\code{yc_brl_with_futures()}, \code{yc_usd_with_futures()}, and \code{yc_ipca_with_futures()} utilize information
from Reference Rates (\code{b3-reference-rates}) and Futures Settlement Prices
(\code{b3-futures-settlement-prices}) datasets to construct a yield curve dataset. This dataset
highlights key vertices and their corresponding underlying futures contracts, providing a
detailed view of the term structure of interest rate.
To prevent excessive memory usage and potential performance issues, these functions include a \code{refdate}
argument that allows filtering the data at the source.
}
\examples{
\dontrun{
date <- preceding(Sys.Date() - 1, "Brazil/ANBIMA")
bova_options <- cotahist_get_options_by_symbols("BOVA11") |> filter(refdate == date)
petr_options <- cotahist_get_options_by_symbols(c("PETR4", "PETR3")) |> filter(refdate == date)
}

\dontrun{
date <- preceding(Sys.Date() - 1, "Brazil/ANBIMA")
yc_brl_with_futures(date)
yc_usd_with_futures(date)
yc_ipca_with_futures(date)
}

}
