% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_dataset}
\alias{template_dataset}
\title{Create a dataset for a given template}
\usage{
template_dataset(template, layer = NULL)
}
\arguments{
\item{template}{The template to use to create the dataset.}

\item{layer}{The layer to use to create the dataset.}
}
\value{
An arrow dataset.
}
\description{
This function creates a dataset for a given template.
}
