% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_meta_load}
\alias{template_meta_load}
\title{Load Metadata for a Template Download}
\usage{
template_meta_load(template, ...)
}
\arguments{
\item{template}{An object representing the template. Can be of class \code{character}
(template ID) or \code{template} (template object).}

\item{...}{Additional arguments used in a specific download.}
}
\value{
The metadata associated with the download.
}
\description{
These functions provide methods to load metadata associated with a template and the arguments used to a
specific download.
}
\details{
The \code{download_marketdata()} function returns a meta object that refers to a specific download.
If the meta object does not exist, it is created.
If the specific download has already been performed in the past, a meta file will exist, and
\code{download_marketdata()} will raise an error upon detecting it.
In such cases, the \code{template_meta_load()} function should be used to load the meta object
associated with the existing meta file.
}
\examples{
# Example usage with a template ID
m <- tryCatch(download_marketdata("b3-indexes-composition"), error = function(e) {
  template_meta_load("b3-indexes-composition")
})
read_marketdata(m)

}
