% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_retrieve}
\alias{template_retrieve}
\title{Retrieve a template by its name}
\usage{
template_retrieve(template_name)
}
\arguments{
\item{template_name}{The name identifying the template to retrieve.}
}
\value{
The template associated with the given name.
}
\description{
This function retrieves a template identified by its name.
}
