% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmprAlign.R
\name{tmprAlign}
\alias{tmprAlign}
\title{Automatically remove cross thermister observation difference}
\usage{
tmprAlign(rB3in, metaD, startDate, endDate, varNames, dTPerctile,
  wndSpdPerctile, logID, Reason, showPlot, savePlot, plotType)
}
\arguments{
\item{rB3in}{data frame input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords (e.g. varNames = "tmpwtr")}

\item{dTPerctile}{percentile of temperature variation to determine mixed periods}

\item{wndSpdPerctile}{percentile of windspeed to make assure mixed periods (optional). wndspd values will be used}

\item{logID}{write an operation identifier to the log frames}

\item{Reason}{reason for hte logged changes (character)}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}

\item{plotType}{'All' = allvars on one plot, compare before and after, '1by1' = separate panel of before and after for each vasr}
}
\description{
This function assign NA to the specific dates range and variables
}
\examples{
newDF <- tmp_align(rB3in, startDate, endDate, varNames, dTPerctile, wndSpdPerctile, logID, Reason, showPlot, savePlot)
}
\keyword{Lake_buoy_data}
