% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLog.R
\name{writeLog}
\alias{writeLog}
\title{writes to the rB3 logKey dataframe}
\usage{
writeLog(rB3in, logID, funName, Reason)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{logID}{write an operation identifier to the log frames, default = NA, 'auto' chooses a log number for you, or provide a numerical code}

\item{funName}{name of the operation associated with the log event}

\item{funName}{the reason for the log event}
}
\description{
writes to the rB3 logKey dataframe
}
\examples{
writeLog(rB3in, logID, funName, Reason)
}
\keyword{logging}
