% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-yc.R
\name{yc_add_bizdays_column}
\alias{yc_add_bizdays_column}
\title{Add Business Days Column to a DataFrame or Arrow Query}
\usage{
yc_add_bizdays_column(x)
}
\arguments{
\item{x}{A \code{data.frame} or an \code{arrow_dplyr_query} object containing
at least the columns \code{refdate} and \code{cur_days}.}
}
\value{
A \code{tibble} with an additional \code{biz_days} column, which represents
business days computed based on a predefined calendar.
}
\description{
This function adds a \code{biz_days} column to the input dataset by calculating
business days using the \code{bizdays::bizdayse} function. It supports both
standard data frames and \code{arrow_dplyr_query} objects.
}
\examples{
\dontrun{
  df_yc_1 <- yc_brl_get() 
  df <- df_yc_1 |> yc_add_bizdays_column()
}
}
