% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggPlotData.R
\name{aggPlotData}
\alias{aggPlotData}
\title{Aggregate a dataset for plotting large time-series quickly
Wide format in, long format out}
\usage{
aggPlotData(dt_in)
}
\arguments{
\item{dt_in}{data in}
}
\description{
Aggregate a dataset for plotting large time-series quickly
Wide format in, long format out
}
\examples{
aggPlotData(dt_in = myDT)

}
\keyword{plotting}
