% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggTS.R
\name{aggTS}
\alias{aggTS}
\title{Aggregate timeseries - low level function}
\usage{
aggTS(dataIn, timestep, FUN, pullAgg)
}
\arguments{
\item{dataIn}{data frame object}

\item{timestep}{standardise (aggregate) to a common timestep (numeric, in mins, to turn aggregation on) !!}

\item{FUN}{aggregation method; mean, median, sum, min, max, or circular (for averaging direction measurements in degrees)}

\item{pullAgg}{aggregate data from before/on new timestamp ('left'; default), either side of timestamp ('centre'), or on/after timestamp ('right')}
}
\description{
Trim dates and/or aggregate data frame
}
\examples{
stdDF <- rB3stdze(rB3in = rawDF, varNames = wqVars, metaD = metaD, startDate = '2016-07-01 00:00:00', endDate = '2018-06-30 23:45:00', doAgg = TRUE, methodAgg = ctrls$methodAgg, pullAgg = ctrls$pullAgg)


}
\keyword{fileIO}
