% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_na.R
\name{assign_na}
\alias{assign_na}
\title{Assign NA to the specified dates range and variables}
\usage{
assign_na(DF_in, metaD, startDate, endDate, varNames, logID)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}
}
\description{
This function assign NA to the specific dates range and variables
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
