% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-futures.R
\name{code2month}
\alias{code2month}
\title{Get month from maturity code}
\usage{
code2month(x)
}
\arguments{
\item{x}{a character with letters that represent the month of maturity of
futures contracts.}
}
\value{
a vector of integers
}
\description{
Get the corresponding month for the string that represent maturities of
futures contracts.
}
\examples{
code2month(c("F", "G", "H", "J", "K", "M", "N", "Q", "U", "V", "X", "Z"))
code2month(c("JAN", "FEV", "MAR", "NOV", "DEZ"))
}
