% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_marketdata}
\alias{download_marketdata}
\title{Download datasets}
\usage{
download_marketdata(template, cache_folder = cachedir(), do_cache = TRUE, ...)
}
\arguments{
\item{template}{the template name}

\item{cache_folder}{Location of cache folder (default = cachedir())}

\item{do_cache}{a logical indicating if the existing file (previously
downloaded) should be used or replaced.}

\item{...}{aditional arguments}
}
\value{
a string with the file path of downloaded file or \code{NULL} if download
fails.

This function downloads data sets for those templates that specifies a
\code{downloader} attribute.
If \code{dest} is not provided, \code{cache_folder} is used and a file with template
id is saved inside it.
}
\description{
Download datasets for a given template.
}
\examples{
\dontrun{
fname <- download_marketdata("CDIIDI")
}

}
