% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftCorr.R
\name{driftCorr}
\alias{driftCorr}
\title{Apply drift correction}
\usage{
driftCorr(rB3in, startDate, endDate, varNames, lowRef, lowStart, lowEnd,
  highRef, highStart, highEnd, logID, showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{variable to which drift correction will be applied}

\item{lowRef}{a reference/known low value at start of period}

\item{lowStart}{equivalent sensor value at start of period which should = lowRef  (i.e., starting offset)}

\item{lowEnd}{equivalent sensor value at end of period which should = lowRef  (i.e., finishing offset)}

\item{highRef}{a reference/known low value at start of period}

\item{highStart}{equivalent sensor value at start of period which should = lowRef  (i.e., starting offset)}

\item{highEnd}{equivalent sensor value at end of period which should = lowRef  (i.e., finishing offset)}

\item{logID}{assign log ID}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}

\item{plotPath}{plot figure of before and after}

\item{logID}{write an operation identifier to the log frames, default = NA}
}
\description{
change slope and offset over time
}
\examples{
rB3in <- driftCorr(rB3in = myDF,DOmgColName = "DOconc_d00050", DOsatColName = "DOpsat_d00050", TColName = "TmpDOs_d00050")


}
\keyword{transformations}
