% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-marketdata.R
\name{fetch_marketdata}
\alias{fetch_marketdata}
\title{Fetch and process market data}
\usage{
fetch_marketdata(template, ...)
}
\arguments{
\item{template}{A character string specifying the market data template to use}

\item{...}{Named arguments that will be expanded into a grid of all combinations
to fetch data for}
}
\description{
Downloads market data based on a template and parameter combinations, then reads
the data into a database.
}
\details{
This function performs two main steps:
\enumerate{
\item Downloads market data files by creating all combinations of the provided parameters
and calling \code{download_marketdata()} for each combination
\item Processes the downloaded files by reading them into a database using \code{read_marketdata()}
}

Progress indicators are displayed during both steps, and warnings are shown for
combinations that failed to download or produced invalid files.
}
\examples{
\dontrun{
fetch_marketdata("b3-cotahist-yearly", year = 2020:2024)
fetch_marketdata("b3-cotahist-daily", refdate = bizseq("2025-01-01", "2025-03-10", "Brazil/B3"))
fetch_marketdata("b3-reference-rates", refdate = bizseq("2025-01-01", "2025-03-10", "Brazil/B3"),
  curve_name = c("DIC", "DOC", "PRE")
)
}

}
