% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_max.R
\name{filter_max}
\alias{filter_max}
\title{Remove large values}
\usage{
filter_max(DF_in, metaD, maxVal, varNames, startDate, endDate, cndFile,
  logID, plotPath)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{cndFile}{conditoin file path}

\item{logID}{assign log ID}

\item{plotPath}{plot figure of before and after}

\item{minVal}{minimum value permitted}
}
\description{
This function assign NA to the data larger than the assigned maxVal
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
