% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_min.R
\name{filter_min}
\alias{filter_min}
\title{Remove small values}
\usage{
filter_min(DF_in, metaD, minVal, varNames, startDate, endDate, cndFile, logID)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{minVal}{minimum value permitted}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{cndFile}{conditoin file path}

\item{logID}{assign log ID}
}
\description{
This function assign NA to the data smaller than the assigned minVal
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))


}
\keyword{wrangling}
