% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_stall.R
\name{filter_stall}
\alias{filter_stall}
\title{Assign NA to the specified dates range and variables}
\usage{
filter_stall(DF_in, metaD, maxRep, varNames, startDate, endDate, cndFile, logID)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{varNames}{list of variable names or keywords}

\item{startDate}{start date}

\item{endDate}{endDate}
}
\description{
This function assign NA to the specific dates range and variables
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
