% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMinMax.R
\name{filterMinMax}
\alias{filterMinMax}
\title{Remove values outside of specified minimum and maximum bounds}
\usage{
filterMinMax(rB3in, startDate, endDate, varNames, filterMin, filterMax,
  logID, Reason, showPlot, savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{filterMin}{apply min value filter, either numeric (filterMin = <value>), or from 'crtls' (filterMin = TRUE)}

\item{filterMax}{apply max value filter, either numeric (filterMax = <value>), or from 'crtls' (filterMax = TRUE)}

\item{logID}{write an operation identifier to the log frames, default = NA}

\item{Reason}{reason for the logged changes (character)}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
Assigns NA to the selected subset of dates and variables
}
\examples{
new <- filterReps(old, varNames = 'TmpWtr', filterMin = 9, filterMax = 25,logID = 10, showPlot = TRUE)
}
\keyword{wrangling}
