% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRoc.R
\name{filterRoc}
\alias{filterRoc}
\title{Assign NA based on filter values from the "ctrls" data frame}
\usage{
filterRoc(rB3in, startDate, endDate, varNames, maxRoc, logID, showPlot,
  savePlot)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{maxRoc}{apply maximum allowed consecutive identical values filter, numerical or TRUE for from 'crtls'}

\item{logID}{write an operation identifier to the log frames, default = NA}

\item{showPlot}{display figure in plots window (TRUE/FALSE)}

\item{savePlot}{save figure to a path (TRUE/FALSE)}
}
\description{
This function removes vals that exceed rate of change
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
