% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-futures.R
\name{futures_get}
\alias{futures_get}
\title{Get futures prices from trading session settlements page}
\usage{
futures_get(refdate)
}
\arguments{
\item{refdate}{reference date used to obtain futures prices.

If \code{refdate} is not provided the last available date is returned, otherwise
the provided date is used to fetch data.}
}
\value{
\code{data.frame} with futures prices.
}
\description{
Scrape page \url{https://www.b3.com.br/en_us/market-data-and-indices/data-services/market-data/historical-data/derivatives/trading-session-settlements/}
to get futures prices.
}
\examples{
\dontrun{
df <- futures_get("2022-04-18")
}
}
