% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idElToModify.R
\name{idElToModify}
\alias{idElToModify}
\title{Identify elements locations for the specified variable and dates ranges}
\usage{
idElToModify(DF_in, startDate, endDate, varNames)
}
\arguments{
\item{DF_in}{data frame input}

\item{startDate}{start date}

\item{endDate}{end date}

\item{metaD}{metadata list}
}
\description{
This sub-function identifies elements of the data frame that satisfies dates range and/or variable names,
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))

}
\keyword{wrangling}
