% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyWriter.R
\name{lazyWriter}
\alias{lazyWriter}
\title{Simple table writing function}
\usage{
lazyWriter(DF_in, fileName, logName, startDate, endDate, varNames)
}
\arguments{
\item{DF_in}{data frame input}

\item{fileName}{filename (path) that csv file will be saved as}

\item{logName}{filename (path) that Rdata log file will be saved as}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}

\item{metaD}{metadata list}
}
\description{
Write table csv file. Also write logName file in Rdata format.
}
\examples{
newDF <- lazyWriter(DF_in,fileName)
}
\keyword{wrangling}
