% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-marketdata.R
\name{read_marketdata}
\alias{read_marketdata}
\title{Read and parses files delivered by B3}
\usage{
read_marketdata(meta, cache_folder = cachedir())
}
\arguments{
\item{filename}{a string containing a path for the file.}

\item{template}{a string with the template name.}

\item{parse_fields}{a logical indicating if the fields must be parsed.}

\item{do_cache}{Whether to use cache or not (default = TRUE)

The function \code{show_templates} can be used to view the available templates.}
}
\value{
\code{data.frame} of a list of \code{data.frame} containing data parsed from
files.
}
\description{
B3, and previously BMF&Bovespa, used to deliver many files with a diverse
set of valuable data and informations that can be used to study of can
be called of marketdata.
There are files with informations about futures, option, interest
rates, currency rates, bonds and many other subjects.
}
\examples{
\dontrun{
path <- "Indic.txt"
df <- read_marketdata(path, template = "Indic")
path <- "PUWEB.TXT"
df <- read_marketdata(path, template = "PUWEB")
}
}
\seealso{
show_templates display_template
}
