% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-marketdata.R
\name{read_marketdata}
\alias{read_marketdata}
\title{Read and parses files downloaded from B3 website.}
\usage{
read_marketdata(meta, cache_folder = cachedir())
}
\arguments{
\item{meta}{a list with the metadata of the downloaded file.}

\item{cache_folder}{location of cache folder (default = cachedir())}
}
\value{
This function returns a \code{data.frame} with the parsed file data.
}
\description{
B3 provides various files containing useful information about
traded assets on the exchange and over-the-counter (OTC) market.
These files include market data, trading data, and asset registration
data (stocks, derivatives, and indices). This function reads these
files and parses their content according to the specifications
defined in the template.
}
\details{
This function reads the downloaded file according to the specifications
and schema defined in the template.
The schema is located in the \code{fields} field of the template and defines the
type of each column in the created dataset.

The generated datasets are saved in a directory named after the
template and are located within the \code{db} directory, which is
inside the \code{cachedir} directory.
The datasets are saved in the Parquet format.
This way, it is possible to use datasets with the template
directories and perform queries using the \code{arrow} package.

If an error occurs while processing a downloaded file, the file and its metadata are removed.
}
\examples{
\dontrun{
meta <- download_marketdata("b3-cotahist-daily", refdate = as.Date("2024-04-05"))
read_marketdata(meta)
}
}
\seealso{
show_templates display_template

cachedir rb3.cachedir
}
