% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyrB3.R
\name{shinyrB3}
\alias{shinyrB3}
\title{Interactive data visualisation}
\usage{
shinyrB3(rB3in, startDate, endDate)
}
\arguments{
\item{rB3in}{rB3 object to be displayed}

\item{startDate}{start date for plot}

\item{endDate}{end date for plot}
}
\description{
Choose and plot interactively single timeseries of source and quality controlled data. \cr
  Select variable from drop-down, choose to display or ignore source data.  \cr
  PLot can be selected and zoomed by brush. Returns x and y information in response to brush selection. \cr
  Example code is generated for removing (set to NA) the highlighted area of the plot
}
\examples{
shinyrB3(rB3in, endDate = '2018-07-01')

}
\keyword{editing}
\keyword{visual}
