% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper-cotahist.R, R/scraper-yc.R
\name{superdataset}
\alias{superdataset}
\alias{cotahist_get_options_by_symbols}
\alias{yc_brl_superset}
\alias{yc_usd_superset}
\alias{yc_ipca_superset}
\title{Enhanced Dataset Creation}
\usage{
cotahist_get_options_by_symbols(symbols)

yc_brl_superset(yc, fut)

yc_usd_superset(yc, fut)

yc_ipca_superset(yc, fut)
}
\arguments{
\item{symbols}{list of symbols to extract market data from the COTAHIST dataset.

The functions \code{cotahist_equity_options_superset()}, \code{cotahist_funds_options_superset()},
\code{cotahist_index_options_superset()}, and \code{cotahist_options_by_symbol_superset()} use
information from the COTAHIST datasets (\code{b3-cotahist-yearly}, \code{b3-cotahist-monthly},
and \code{b3-cotahist-daily}) and Reference Rates (\code{b3-reference-rates}) and return a dataframe
containing stock option data, including the closing price of the underlying stocks, the ticker
of the underlying asset, and the interest rate at the option's expiration. The returned dataframe
contains the following columns: "refdate", "symbol", "type", "symbol_underlying",
"strike_price", "maturity_date", "r_252", "close", "close_underlying", "volume",
"trade_quantity", and "traded_contracts".

\code{cotahist_options_by_symbol_superset()} returns the same dataset but filtered for the specified asset ticker.}

\item{fut}{futures dataset}
}
\value{
A dataframe with the super-dataset.
}
\description{
To maximize the utility of B3's existing datasets, several functions integrate data from multiple
sources to generate specialized datasets for specific analytical needs. For instance,
\code{cotahist_equity_options_superset()} combines data from COTAHIST datasets
(\code{b3-cotahist-yearly}, \code{b3-cotahist-monthly}, and \code{b3-cotahist-daily}) and Reference Rates
(\code{b3-reference-rates}) to construct a dataset containing stock options data. This dataset
includes details such as the closing price of the underlying stock, its ticker symbol, and the
applicable interest rate at option expiration. This comprehensive data enables users to perform
option pricing and calculate implied volatility.
}
\details{
\code{yc_brl_superset()}, \code{yc_usd_superset()}, and \code{yc_ipca_superset()} utilize information
from Reference Rates (\code{b3-reference-rates}) and Futures Settlement Prices
(\code{b3-futures-settlement-prices}) datasets to construct a yield curve dataset. This dataset
highlights key vertices and their corresponding underlying futures contracts, providing a
detailed view of the term structure of interest rate.
}
\examples{
\dontrun{
date <- Sys.Date() - 1
bova_options <- cotahist_get_options_by_symbols("BOVA11") |> filter(refdate == date)
petr_options <- cotahist_get_options_by_symbols(c("PETR4", "PETR3")) |> filter(refdate == date)
}

\dontrun{
date_ <- preceding(Sys.Date() - 1, "Brazil/ANBIMA")
fut <- futures_get() |> filter(refdate == date_)
yc <- yc_brl_get() |> filter(refdate == date_)
yc_superset(yc, fut)
}

}
