% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_align.R
\name{tmp_align}
\alias{tmp_align}
\title{Automatically remove cross thermister observation difference}
\usage{
tmp_align(DF_in, metaD, startDate, endDate, varNames, dTPerctile,
  wndSpdPerctile, logID, plotPath)
}
\arguments{
\item{DF_in}{data frame input}

\item{metaD}{metadata list}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{varNames}{list of variable names or keywords}
}
\description{
This function assign NA to the specific dates range and variables
}
\examples{
newDF <- exclude_vars(myDF,metaData,varNames = c("pH","wndDir"))
}
\keyword{wrangling}
