% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varWrangle.R
\name{varWrangle}
\alias{varWrangle}
\title{Interpolate the date range given for the variables specified}
\usage{
varWrangle(rB3in, varNames, task)
}
\arguments{
\item{rB3in}{rB3 object input}

\item{varNames}{list of variable names or keywords}

\item{task}{one of: rm, add, or moveto}

\item{loc}{N: move or add varNames after original Nth variable}
}
\description{
Linearly interpolates gaps (NAs) within the selected subset of dates and variables
}
\examples{
newDF <- varWrangle(rB3in, varNames = "tmpWtr", task = "sort")
newDF <- varWrangle(rB3in, varNames = "tmpWtr", task = "moveto", loc = 3)
newDF <- varWrangle(rB3in, varNames = "DOSat", task = "add")
newDF <- varWrangle(rB3in, varNames = "tmpWtr", task = "rm")
}
\keyword{wrangling}
