% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLAinputs.R
\name{writeLAinputs}
\alias{writeLAinputs}
\title{Write input files for Lake Analyzer from your rB3 files}
\usage{
writeLAinputs(rB3in, startDate, endDate, wtrNames, wndName, wndHeight,
  filePath)
}
\arguments{
\item{rB3in}{data frame to be standardized}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{wtrNames}{variables (columns) of water temperature to be written to .wtr file}

\item{wndName}{variable (column) of wind speed data to be written to .wnd file}

\item{wndHeight}{height of wind speed sensor above lake surface}

\item{filePath}{folder to which files will be writen, set relative to working dir}
}
\description{
Easily generate Lake Analyzer water temperature (.wtr) and wind speed (.wnd) files with the correct header formats
}
\examples{
writeLAinputs(rB3in = rB3agg2, wtrNames = 'TmpWtr', wndName = 'WndSpd', wndHeight = 1.5)


}
\keyword{fileIO}
