% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLAinputs.R
\name{writeLAinputs}
\alias{writeLAinputs}
\title{Write input files for Lake Analyzer from your rB3 files}
\usage{
writeLAinputs(rB3in, startDate, endDate, wtrNames, wndName, wndHeight,
  filePath)
}
\arguments{
\item{rB3in}{data frame to be standardized}

\item{startDate}{start date}

\item{endDate}{endDate}

\item{wtrNames}{variables (columns) of water temperature to be written to .wtr file}

\item{wndName}{variable (column) containing wind speed data .wnd file}

\item{wndHeight}{height of wind sensor (in m) above water surface, used to stadardise wind speed to 10 m height using wind power law}

\item{filePath}{folder to which files will be writen, set relative to working dir}
}
\description{
Easily generate Lake Analyzer water temperature (.wtr) and wind speed (.wnd) files with the correct header formats
}
\examples{
stdDF <- rB3stdze(df = rawDF, varNames = wqVars, startDate = '2016-07-01 00:00:00', endDate = '2018-06-30 23:45:00', doAgg = TRUE, methodAgg = ctrls$methodAgg, pullAgg = ctrls$pullAgg)


}
\keyword{fileIO}
